// snapshots/snapshot-2024-10-30T19-52-31-242Z/src/types.ts
export type AnimationDirection = 'center' | 'left' | 'right' | 'top' | 'bottom'
export type AnimationType = 'none' | 'linear' | 'bounce' | 'elastic-out'
export type ShapeType = 'rectangle' | 'circle' | 'triangle' | 'pentagon' | 'star' | 'speech-bubble' | 'camera' | 'image'

export interface Animation {
  type: AnimationType
  direction: AnimationDirection
  duration: number
}

export interface KeyframeAnimation {
  type: AnimationType
  duration?: number
}

export interface Shape {
  id: number
  type: ShapeType
  x: number
  y: number
  width: number
  height: number
  rotation: number
  scale: number
  color: string
  keyframes: Keyframe[]
  depth: number
  image?: string
  startFrame?: number
  endFrame?: number
  animation: {
    enter: Animation
    exit: Animation
  }
}

export interface Keyframe {
  frame: number
  x: number
  y: number
  rotation: number
  scale: number
  depth: number
  zoom?: number
  animation?: KeyframeAnimation
}